Add files via upload
Hi All,
this is a Open source design of a ventilator (Ambu) bag inflator without any programming or Arduino boards. this design is inspired from MIT design, but uses a readily available windshield motor, so that Auto companies can mass manufacture for this Pandemic. This design uses only a Speed regulator and all details are in the PDF and STEP files for anyone to quickly manufacture. This design was done with the help of professionals from Applied Materials India on their individual capacity. This design is not tested and the designers take no responsibility in the expected results.